requirements.txt
# vNAS Sweatbox Scenario Generator Requirements

# Core dependencies
requests>=2.31.0

# GUI dependencies
Pillow>=10.0.0

# Optional: for better performance with large JSON files
# ujson>=5.8.0
